requirements.txt
tzdata
Django==4.0.6
djangorestframework==3.13.1
